requirements.txt
astroid==2.15.6
beautifulsoup4==4.12.2
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
filelock==3.12.2
identify==2.5.26
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysql-connector-python==8.1.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
protobuf==4.21.12
pylint==2.17.5
pytest==7.4.0
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.11.0
six==1.16.0
soupsieve==2.4.1
tomlkit==0.12.1
urllib3==2.0.4
virtualenv==20.24.3
wrapt==1.15.0
